+++ thehive/app/org/thp/thehive/models/CaseTemplate.scala
@@ -14,14 +14,14 @@ case class CaseTemplateCustomField(
     stringValue: Option[String] = None,
     booleanValue: Option[Boolean] = None,
     integerValue: Option[Int] = None,
-    floatValue: Option[Float] = None,
+    floatValue: Option[Double] = None,
     dateValue: Option[Date] = None
 ) extends CustomFieldValue[CaseTemplateCustomField] {
   override def order_=(value: Option[Int]): CaseTemplateCustomField            = copy(order = value)
   override def stringValue_=(value: Option[String]): CaseTemplateCustomField   = copy(stringValue = value)
   override def booleanValue_=(value: Option[Boolean]): CaseTemplateCustomField = copy(booleanValue = value)
   override def integerValue_=(value: Option[Int]): CaseTemplateCustomField     = copy(integerValue = value)
-  override def floatValue_=(value: Option[Float]): CaseTemplateCustomField     = copy(floatValue = value)
+  override def floatValue_=(value: Option[Double]): CaseTemplateCustomField    = copy(floatValue = value)
   override def dateValue_=(value: Option[Date]): CaseTemplateCustomField       = copy(dateValue = value)
 }
